Move latex_functions.rs content into latex_constructs.rs (Perl parity)
In Perl, appendix helpers, note tags, footnote relocation, preamble check,
and tabular_bindings are all inline in latex_constructs.pool.ltxml.
Eliminates the Rust-only latex_functions.rs file.

413 tests pass.
